% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/A_edat2.R
\docType{data}
\name{A_edat2}
\alias{A_edat2}
\title{Data: An example e-prime output file from Intentional Binding pardigm}
\description{
Representative data from healthy control subject for trying out the package functions.
This data in Rda format can be located in data subfolder and in raw tab-delimited text format in extdata subfolder of Intentional Binding package directory.
}
\author{
Transpsych Lab
}
\references{
\url{www.transpsychlab.org}
}
\keyword{data}

