% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/C_edat2.R
\docType{data}
\name{C_edat2}
\alias{C_edat2}
\title{Data: An example e-prime output file from Intentional Binding pardigm}
\description{
Representative data for trying out the package functions.
This data in raw format can be located in extdata subfolder of Intentional Binding package directory.
}
\author{
Transpsych Lab
}
\references{
\url{www.transpsychlab.org}
}
\keyword{data}

