% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/completeFun.R
\name{completeFun}
\alias{completeFun}
\title{completeFun}
\usage{
completeFun(dat, desiredCols)
}
\arguments{
\item{dat}{A data frame}

\item{desiredCols}{A vector of column names}
}
\value{
Data frame with complete rows
}
\description{
Extract rows with complete data
}
\details{
This is a helper function to eliminate trials missing critical data such as key press.
}
\author{
Sunil V Kalmady (Citation: stackoverflow.com)
}
\references{
\url{www.transpsychlab.org}
}

