% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBreport.R
\name{IBreport}
\alias{IBreport}
\title{IBreport}
\usage{
IBreport(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate detailed markdown HTML report from Intentional Binding summary data for single subject.
}
\details{
IBsummary should be run first and SummaryData.csv should be available.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBreport("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding report for single subject
IBsummary("A.edat2.txt")
IBreport("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

