% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBreport_multi.R
\name{IBreport_multi}
\alias{IBreport_multi}
\title{IBreport_multi}
\usage{
IBreport_multi()
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate detailed HTML markdown report from Intentional Binding summary data for multiple subjects.
}
\details{
IBsummary should be run first and more than one SummaryData.csv files should be available.
Table of combined dataset is also generated.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBsummary("Brain2010-IB-0_50_75-H-3-1.edat2.txt")
IBreport_multi()

# Try out the example data provided with the package
# copy multiple txt files - A.edat2.txt and H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/", package = "IntentionalBinding")
flist <- list.files(srcpath, ".txt$", full.names = TRUE)
destpath <- getwd()
file.copy(flist, destpath)

#Run multisubject Intentional binding report
IBsummary("A.edat2.txt")
IBsummary("H.edat2.txt")
IBreport_multi()
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

