% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBresult.R
\name{IBresult}
\alias{IBresult}
\title{IBresult}
\usage{
IBresult(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate ANOVA results for Intentional Binding summary data.
}
\details{
IBsummary should be run first and SummaryData.csv should be available.
Output is only displayed on the console.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBresult("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding result function
IBsummary("A.edat2.txt")
IBresult("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

