% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBresultplot.R
\name{IBresultplot}
\alias{IBresultplot}
\title{IBresultplot}
\usage{
IBresultplot(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate box-plots for results from Intentional Binding summary data.
}
\details{
IBsummary should be run first and SummaryData.csv should be available.
Box Plots are output as two pdf files. One with all the clean trials, another after removal of outliers.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBresultplot("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding condition wise box-plots
IBsummary("A.edat2.txt")
IBresultplot("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

