% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBresulttable.R
\name{IBresulttable}
\alias{IBresulttable}
\title{IBresulttable}
\usage{
IBresulttable(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate result tables with summary statistics for Intentional Binding summary data.
}
\details{
IBsummary should be run first and SummaryData.csv should be available.
Two tables are generated with mean, N, SD, CV for each experimental condition. One with all the clean trials, another after removal of outliers.
Predictive and retrospective components of intentional binding are also generated.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBresulttable("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding results table
IBsummary("A.edat2.txt")
IBresulttable("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

