% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBstepplot.R
\name{IBstepplot}
\alias{IBstepplot}
\title{IBstepplot}
\usage{
IBstepplot(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate plots of step periods for each trial in Intentional Binding summary data and save into single pdf with multiple pages.
}
\details{
IBsummary should be run first and SummaryData.csv should be available.
This will be helpful in detecting any timing issues during administration of experiment.
Plots are annotated with critical time points such as key press.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")
IBstepplot("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding step duration plots
IBsummary("A.edat2.txt")
IBstepplot("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

