% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IBsummary.R
\name{IBsummary}
\alias{IBsummary}
\title{IBsummary}
\usage{
IBsummary(fname)
}
\arguments{
\item{fname}{Filename of unicode tab-delimited text from e-prime experiment output.}
}
\description{
Generate critical metrics and trial wise summary for raw eprime data of Intentional Binding paradigm.
}
\details{
Generates SummaryData.csv which is critical for other functions in package which carryout downstream data processing.
First, the function strips off trials without keypress, subject judgement or audio failure.
Then, prepares table with clock positions and their timings in order to reconstruct every trial.
Timing corresponding to judged hand position is obtained. This is done by linear interpolation of timing corresponding to judged hand position for six rotations and then selecting the timepoint closest to key press.
Additional Columns with the first trial and outliers removed [based on tukey's criterion] as well as average clock step period per trial are appended.
These values along with other critical parameters obtained from raw eprime data is output as single CSV table.
}
\examples{
\dontrun{
# A general example
IBsummary("Brain2010-IB-0_50_75-A-3-1.edat2.txt")

# Try out the example data provided with the package
# copy A.edat2.txt or H.edat2.txt to working directory (manually or using codes below)
srcpath <- system.file("extdata/A.edat2.txt", package = "IntentionalBinding")
destpath <- getwd()
file.copy(srcpath, destpath)

#Run Intentional binding summary function
IBsummary("A.edat2.txt")
}
}
\author{
Sunil V Kalmady
}
\references{
\url{www.transpsychlab.org}
}

