% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeOut.R
\name{removeOut}
\alias{removeOut}
\title{removeOut}
\usage{
removeOut(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{A boolean that indicates whether to ignore NA's}
}
\value{
Vector with outliers removed
}
\description{
Remove outliers using tukey's method
}
\details{
This is a helper function to eliminate trials with outlying judgment values.
}
\author{
Sunil V Kalmady (Citation: stackoverflow.com)
}
\references{
\url{www.transpsychlab.org}
}

