% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{summarySE}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{measurevar}{The name of a column that contains the variable to be summarized}

\item{groupvars}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is 95)}
}
\value{
Dataframe with descriptive statistics
}
\description{
Gives count, mean, standard deviation, standard error of the mean, coefficient of variation and confidence interval.
}
\details{
This is a helper function to obtain descriptives and coefficient of variation.
}
\author{
Sunil V Kalmady (citation: cookbook-r.com)
}
\references{
\url{www.transpsychlab.org}
}

